+++ app/(site)/components/AboutSection/LanguageSection.tsx
@@ -1,10 +1,10 @@
 import Tag from "@/components/Atoms/Tag";
+import LanguageModal from "@/components/Modal/LanguageModal";
 import HeadingThree from "@/components/Text/HeadingThree";
 import { languages } from "@/constants/languages";
 import { Repository } from "@/types/languages";
 import { Skill } from "@/types/skills";
 import React, { useState } from "react";
-import LanguageModal from "../../Modal/LanguageModal";
 
 /**
  * Displays a list of languages that I know.
